Basic poll support: title, choice count, and all the standard fields
Showing just the number of poll options is simplest, since the HN API
includes that value directly (as "parts").

Getting the label for each poll option would make the output quite a bit
longer at the cost of one additional API request per choice. (That seems
excessive, so I'm not doing it. At least not yet.)